Separate compilation of kernel and app.
The two are compiled separately and then linked by linker scripts that
agree on the kernel-app boundaries in ROM and RAM.  These are currently
hardcoded.

By loading one after the other in GDB, we place them in their respective
Flash areas.  Unfortunately GDB doesn't remember the symbols from the
second load, contrary to its docs, so we have to take an additional step
and introduce the symbol files at explicit addresses.