k/address_range: new memory access model.
The application's kernel configuration now specifies two or more ranges
of memory with attributes.  The system can mint keys into this region
that act as memory region keys, and can (among other things) be loaded
into the region registers of a Context.

This frees the application from needing to work exclusively in terms of
power-of-two regions, because address ranges can be arbitrary.  It makes
holding a Context key far less scary, because to affect the memory map,
one must also hold a legitimate memory region key.

The protocol for memory region keys allows the derivation of new keys
with less authority, the inspection of the hardware settings that would
result from loading the key into a region register, and that's it.  In
the future we might allow data access through the keys, but it isn't
implemented.

This commit includes the first batch of unit tests for primitive kernel
objects, which is exciting, if you're me.