k: change messaging protocol to reduce overhead.
This reduces the number of calls, and particularly virtual calls,
required to deliver a message inside the kernel.  Per-byte overhead in
the demo application is down by 12%.

The new protocol clarifies the role of priority in message delivery,
eliminating a whole mess of TODOs around the priority choices for reply
senders.

Overall LOC down by 2.15%.