k/{maybe, range_ptr}: alias ETL types.
This means (1) we can use shorter names to refer to these types, and (2)
we can choose the check policy used in both.  Here the config::checks
constant determines whether uses of Maybe and RangePtr are guarded by
asserts or not.  (Default: guarded.  It's not terribly expensive and
will help to catch bugs faster.)