k/context: improve on_delivery_accepted performance.
This small change contains two subtle fixes.

First, it reuses the reply key generated earlier in the call, rather
than producing a new one (not terribly expensive, but not free).

Second, it uses a reference to the source key instead of a copy,
eliminating a four-word transfer.  This also has the (positive) side
effect of revoking the receive-side key if it's been invalidated,
whereas the original implementation would revoke the *copy.*