k/scheduler: elide maybe check if redundant.
Originally I had imagined that we'd remove this assert completely once
Maybe checking was turned on ... but given that I've allowed for
checking to be disabled, and this condition could be violated by the
behavior of the system (not a kernel bug), we need to keep this assert.

So the least we can do is disable it when Maybe checks are on.