k: move SysTick to separate object.
This reverses my recent change to add SysTick support to Interrupt, and
instead models SysTick as a separate object.  This is a result of my
discovery that I can't use the MPU to grant unprivileged access to PPB
peripherals such as SysTick -- the kernel must proxy it.

I've refactored Interrupt to try and minimize code duplication.