k/context: use right MPU settings to write registers.
Whoops.  When reading registers from a Context's stack, we correctly
swap out the MPU configuration so that the call still works if the
caller doesn't have direct access to the target's stack.

However, when writing registers, I forgot to include the equivalent
code.

So setting a Context's registers will only work when the caller can see
into the target's stack.  This is common enough (e.g. omniscient System
task) that I didn't notice the behavior; I happened to notice this when
golfing the code.