k/interrupt: fix error in message key production.
The Interrupt object (and by extension SysTick) was helpfully trying to
provide its target with a key that could be used to re-enable the
interrupt.

And then it was nulling the key out on the next line.  Whoops.

This code is pretty ancient; I finally noticed this because previous
applications have held a key to the Interrupt object that they used to
re-enable, rather than using the key delivered with the message.