k/interrupt: handle unexpected repeated interrupts.
This is a condition the application can trigger, without requiring a
hardware or kernel bug.  If the application re-enables the interrupt
from a task without receiving the interrupt message, the next interrupt
won't have anywhere to go.

There are a few alternate ways of solving this.  One would be to have
the 'enable' IPC operation withdraw the interrupt sender from any block
list that it's a part of.

All that being said: I've caught the SysTick timer generating double
interrupts.  I'm having a hard time reproducing it, and it's almost
certainly something I'm doing wrong, but this keeps the kernel from
asserting when it occurs.