k/context: add context save/restore ops.
These were inspired by the FreeRTOS port, which swaps FreeRTOS tasks
through a single kernel Context.  This was generating a lot of RPCs.

The idea of barfing all the kernel-saved registers somewhere that they
can be inspected feels quite general to me.  It's useful for
task-swapping, yes, but also potentially for debugging and
introspection.