k: make send virtual protocol cheaper, safer.
Two operations in the send protocol, on_delivery_accepted and
on_blocked_delivery_accepted, need a Message buffer in which to deposit
the sent message.

Previously, this was passed in by reference as an out-parameter.  But
since Message's constructor does the right thing and zeroes its memory,
this meant that we were zeroing message buffers just to fill them in.

On the other hand, we don't want to jump through hoops to leave the
buffer *uninitialized*, because there's a risk someone might try to use
it.

The solution is to return the Message instead.  This is the
language-supported method for passing in an uninitialized buffer and
making sure it gets initialized before return.

Unfortunately we can only do this with one buffer, so Keys is still an
out-parameter for now.  I'll evaluate packing the two into one struct
later.