k/address_range: remove attributes
After consideration, I've decided that the inherent read-only/XN
attributes on AddressRange don't mean much to the kernel.  Because even
a read-only region can't alias kernel memory, it's no skin off the
kernel's metaphorical back if the application writes to the region.
It might fault, sure, but that's not a problem.

The useful implication of this is that AddressRange objects are now
quite small.

This change is partial, because I'm holding off on breaking the ABI
until the rest of the changes I'm working on break it quite thoroughly.