k/memory: redesign for new model.
Memory objects are now of fixed extent, and keys do not have the power
to describe a subset of the object.

Keys now express the complete set of ARMv7-M memory attributes in an
unencoded form.  Applications now have slightly more rope.

Like a lot of my commits lately, this is a transitional form.  The
object map still exists, the kernel still sets up a bunch of objects on
the app's behalf, and those objects still come from "donated RAM"
instead of memory objects.