Change how Context and ReplyGate are introduced.
Programs now use a key to designate the ReplyGate for a Context, which
is good.

To prevent them from using a key delivered with a message, the ReplyGate
now has a notion of whether it is "bound" to a Context.  The key given
at creation must not yet be bound.