k/context: no access to stacked registers.
The register access operations on Context have been redundant for a
while.  This change is an improvement: it takes away the ability to
access the registers saved on the stack by hardware.  This is one fewer
IPC path that might fault, and saves quite a few repetitive lines of
code.

Systems are expected to access stacked registers either directly, or by
messaging the Memory object containing the stack in question.

The more extreme version of this change would remove the single-register
access calls, but currently that would leave me without a way to access
the stack pointer.  Needs more thought.