ObjectTable: add get_kind.
I debated where to place this operation.  EROS and kin do it by a
message to the object being scrutinized ... which lets objects lie about
their natures, something I approve of.

However.  I don't currently *have* any messages that are understood by
all objects.  In particular, Gate and ReplyGate are completely
transparent.  So it's not clear how we'd tell *their* kind, as opposed
to the kind claimed by the program serving the Gate.

Moreover, the mechanism used by EROS and kin does not appear to be
cleanly extensible.

So I've done my usual thing and *totally punted.*  This does not
introduce a way to query an arbitrary object for its kind, nor any
extensible kind-code representation.  Rather, I'm treating it as a
reflective operation, and placing it in the natural place for such
operations: on ObjectTable.