Context: re-golf IPC dispatch.
I was actually spending more lines of code on the function prototypes
and definitions for the do_foo functions than on the actual code!
That's a sign.

I've flattened all the method implementations into the dispatch switch
statement.  This does not, IMHO, make it unwieldy yet, and saves a
tremendous amount of boilerplate (-96 SLOC net).