Expose key maps in IPC ABI.
The assignment of message keys to key registers is still hardcoded, only
now it's hardcoded in the demo library instead of the kernel.

I had to move the IPC stub into assembly, because (even with the naked
attribute) GCC was generating a prologue-epilogue that trashed the third
function parameter in r3.  Something having to do with struct return,
from what I can tell.  Oh well.