Replace Reply Gates with Context reply keys.
Pleasant effects include:
- Number of kernel object kinds reduced by 1.
- Weird attachment of Context and Reply Gate mooted.
- Reply path slightly faster.
- One fewer preallocated object.
- Precedent for distinguishing transparent vs. service Gate keys.

This change has a lot of small implications, but the meat of it is in
Context.

Note that this reduces the "call count" used to distinguish successively
issued reply keys from 96 bits to 63 bits.  Assuming a call per
nanosecond (currently very optimistic) this gives a rollover period of
over 292 years.