Kernel-side invalidation support.
Adds a new Invalidate method to ObjectTable, which makes all outstanding
keys to an object invalid.  (The caller is expected to use its leet
ObjectTable powers to make a new key.)

This implies object-specific behaviors to invalidate cases in the kernel
where I use direct pointers instead of keys, for performance reasons.
So I've added Object::invalidate_hook in support of that.