k/panic: don't be weak.
The original idea here was good -- allowing applications to add their
own panic handlers without needing to recompile the kernel -- but I did
it wrong.  There needs to be a strong-linkage panic handler and a weak
alias.  For now, I'm just removing the weak attribute, for it is causing
me trouble.