k/testutil/panic_fake: don't FAIL.
Sometimes it's useful, in a test, to assert that an operation panics.  I
could abort() here and use a death test, but they're relatively
expensive.  It's better to throw a predictable exception and use
ASSERT_THROW.

There is no ASSERT_FAILS.  The FAIL() macro here was interfering with my
ability to sense the std::logic_error.

This results in a very small loss of clarity for test output (the
message "PANIC:" no longer appears).