k/context: update MPU more carefully.
So, now that I'm actually applying MPU settings on context switch (oops)
I found a problem.

If an MPU region is enabled, there is no obvious safe order for updating
RBAR/RASR.  In my test application, an update managed to create a large
XN region covering the kernel, leading to an immediate fault.

I'll revisit how to do this with finesse later, but for now, I'm doing
the conservative thing and disabling each region before update.