a/rt/keys: add reserve_keys.
This allows multiple keys to be reserved in one shot, potentially
reducing the size of small programs with fixed key registers (i.e. all
of them, right now).

Initially this actually *increases* the size of the binaries, because
of the additional assert and its corresponding string table entries.
Working on it.