a/srv/ascii: turn on LTO.
This improves GCC's ability to deduplicate entries in both the GOT and
the string table, besides enabling more general optimizations.

The ASCII server binary is 49% smaller after this change, mostly because
the optimizer noticed that no assert strings are actually used, and
cleaned up the string table.  I'm about to undo some of that, because
assert messages are useful.