Shift kernel selectors to avoid zero.
This resolves a long-standing TODO.

Early on, I numbered method selectors starting at zero.  I quickly
realized that reply messages were likely to use selector zero, and that
using it for both purposes raised the risk of a reply message being
misinterpreted as a method call, if somebody sent a real key in k0.

This makes the convention uniform: selector zero is only used for
replies.