Refactor Context state changes, reduce duplication.
This reclaims the Context method 'make_runnable' to perform a common
series of state updates, moving its former logic into the message decode
switch.  This makes it harder for any of these routines to diverge.

While I was in there, I made it clearer when context switches get
pended, and was able to remove some code from the startup sequence.